Time: 15 ms (63.22%), Space: 15.5 MB (97.76%) - LeetHub

+++ 0167-two-sum-ii-input-array-is-sorted/0167-two-sum-ii-input-array-is-sorted.cpp
@@ -8,11 +8,11 @@ public:
         vector<int> ans(2,1);
         while(i<j){
             
-            if((long long)numbers[i]+(long long)numbers[j]==target){
+            if(numbers[i]+numbers[j]==target){
                 ans[0]+=i;
                 ans[1]+=j;
                 return ans;
-            }else if((long long)numbers[i]+(long long)numbers[j]>target){
+            }else if(numbers[i]+numbers[j]>target){
                 j--;
             }else{
                 i++;
